Move tests to own dir; fix flask run
