+++ src/app.py
@@ -13,4 +13,4 @@ def create_app(config_name):
     db.init_app(app)
 
     return app
-
+
